Moving AuditService to datastore folder
