+++ src/main/java/cloud/fogbow/fns/Main.java
@@ -14,7 +14,7 @@ import cloud.fogbow.fns.constants.ConfigurationPropertyKeys;
 import cloud.fogbow.fns.core.datastore.DatabaseManager;
 import cloud.fogbow.fns.core.datastore.orderstorage.RecoveryService;
 import org.apache.log4j.Logger;
-import cloud.fogbow.fns.core.datastore.orderstorage.AuditService;
+import cloud.fogbow.fns.core.datastore.AuditService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
