+++ src/main/java/cloud/fogbow/fns/core/datastore/DatabaseManager.java
@@ -2,7 +2,6 @@ package cloud.fogbow.fns.core.datastore;
 
 import cloud.fogbow.fns.core.datastore.orderstorage.RecoveryService;
 import org.apache.log4j.Logger;
-import cloud.fogbow.fns.core.datastore.orderstorage.AuditService;
 import cloud.fogbow.fns.core.model.FederatedNetworkOrder;
 import org.springframework.beans.factory.annotation.Autowired;
